Created new file for utility stuff; added DetailsIntent; added missing mensas to interaction model
